requirements.txt
alembic==1.7.5
amqp==5.0.9
APScheduler==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
dnspython==2.1.0
defusedxml==0.7.1
dominate==2.6.0
email-validator==1.1.3
Flask==2.0.2
Flask-APScheduler==1.12.3
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
httpie==2.6.0
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.2
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.2
PyJWT==2.3.0
PySocks==1.7.1
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-toolbelt==0.9.1
soupsieve==2.3.1
SQLAlchemy==1.4.29
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vine==5.0.0
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.1
yahoofinancials==1.6
zipp==3.6.0